refactor(B2CQA-5582): split quick-amount suite and dedupe per Bruno's review
- Move the Max/Balance/% quick-fill suite out of entrypoint.swap.spec.ts
  into its own quickAmount.swap.spec.ts, following the sibling
  <feature>.swap.spec.ts convention (no shared registration needed,
  playwright.config.ts auto-discovers *.spec.ts under tests/specs).
- Add SwapPage.checkPercentageButtonsEnabled to replace the 3 repeated
  isPercentageEnabled("25%"/"50%"/"75%") assertion blocks.
- Add SwapPage.checkPercentageFillsBalance to move the proportional-
  amount math out of the spec and into a single @step-decorated method,
  which also removes the redundant test.step wrapper that was
  double-nesting under clickPercentage's own @step in Allure.
- Extract expectAmountCloseTo/parseBalanceAmount into a new
  amountUtils.ts (no tests/page dependency) so swap.page.ts can import
  them without a swapUtils.ts -> tests/page -> swap.page.ts cycle.
