+++ e2e/desktop/tests/utils/swapUtils.ts
@@ -17,7 +17,6 @@ import {
 import { getEnv } from "@ledgerhq/live-env";
 import * as allure from "allure-js-commons";
 import BigNumber from "bignumber.js";
-import { expect } from "@playwright/test";
 import { launchSpeculos, cleanSpeculos } from "./speculosUtils";
 
 export function setupEnv(disableBroadcast: boolean = false): void {
@@ -27,21 +26,6 @@ export function setupEnv(disableBroadcast: boolean = false): void {
       ...(disableBroadcast ? { DISABLE_TRANSACTION_BROADCAST: "1" } : {}),
     },
   });
-}
-
-export function expectAmountCloseTo(actual: number, expected: number, relativeTolerance = 0.01) {
-  const tolerance = Math.max(expected * relativeTolerance, 1e-6);
-  expect(Math.abs(actual - expected)).toBeLessThan(tolerance);
-}
-
-// Balance labels are prefixed with a translated label (e.g. "Balance 1,234.56 ETH"),
-// so a plain parseFloat would return NaN; extract the numeric substring and strip
-// locale thousands separators before parsing.
-export function parseBalanceAmount(balanceText: string | null): number {
-  const numericMatch = balanceText?.match(/\d[\d,]*(?:\.\d+)?/)?.[0];
-  const balance = numericMatch ? Number.parseFloat(numericMatch.replaceAll(",", "")) : Number.NaN;
-  expect(Number.isFinite(balance)).toBe(true);
-  return balance;
 }
 
 export async function checkAccountFromIsSynchronised(app: Application, swap: Swap) {
